refactor(domain): reorganize value_objects into package structure
Convert monolithic value_objects.py module into a semantic package:
- identifiers.py: RunId, StageName, EntityName, PipelineId, ChemblId
- crypto.py: HashDigest
- network.py: HttpUrl
- temporal.py: Timestamp

All classes re-exported from __init__.py for backward compatibility.
